+++ src/products/imp/CompositeProduct.h
@@ -35,7 +35,7 @@ public:
 
     void addPart(Product *product);
 
-    int getCurrentPrice() const override;
+    double getCurrentPrice() const override;
 
     std::string getType() const override;
