+++ src/products/imp/Display.cpp
@@ -19,7 +19,7 @@ Display::~Display() {
 
 }
 
-int Display::getCurrentPrice() const {
+double Display::getCurrentPrice() const {
     int ageInDays = getAge();
     return ageInDays < 30 ? initPrice : (int) initPrice * (ageInDays >= 30 && ageInDays < 90 ? 0.9 : 0.8);
 }
