+++ src/products/imp/HardDisk.cpp
@@ -4,7 +4,7 @@
 
 #include "HardDisk.h"
 
-int HardDisk::getCurrentPrice() const {
+double HardDisk::getCurrentPrice() const {
     int ageInDays = getAge();
     return ageInDays < 30 ? initPrice : (int) initPrice * (ageInDays >= 30 && ageInDays < 90 ? 0.9 : 0.8);
 }
