+++ src/products/imp/VideoCard.h
@@ -27,7 +27,7 @@ public:
 
     virtual ~VideoCard();
 
-    int getCurrentPrice() const override;
+    double getCurrentPrice() const override;
 
     std::string getType() const override;
